+++ samples/Sample_03_Sections.php
@@ -16,8 +16,8 @@ $section->addPageBreak();
 $section->addPageBreak();
 
 // New portrait section
-$section = $phpWord->addSection(array('marginLeft' => 600, 'marginRight' => 600, 'marginTop' => 600, 'marginBottom' => 600));
-$section->addText('This section uses other margins.');
+$section = $phpWord->addSection(array('paperSize' => 'Folio', 'marginLeft' => 600, 'marginRight' => 600, 'marginTop' => 600, 'marginBottom' => 600));
+$section->addText('This section uses other margins with folio papersize.');
 
 // New portrait section with Header & Footer
 $section = $phpWord->addSection(array('marginLeft' => 200, 'marginRight' => 200, 'marginTop' => 200, 'marginBottom' => 200, 'headerHeight' => 50, 'footerHeight' => 50,));